refactor(terminal-protocol): graduate encodeKey + ACCEPTED_KEY_NAMES from kaval-tui
The kolu MCP face's lifecycle_sendInput speaks the same named-key vocabulary
(Enter/Escape/chords as separate sends), making it the second verbatim
consumer of kaval-tui's key encoder — the exact birth condition of
@kolu/terminal-protocol. Move encodeKey + the ACCEPTED_KEY_NAMES drift-guarded
help string (and their tests) beside the NAMED_KEY_BYTES table they fold
through; kaval-tui imports them back.
