+++ packages/kaval-tui/src/send.test.ts
@@ -1,8 +1,5 @@
-import { NAMED_KEY_BYTES } from "@kolu/terminal-protocol";
 import { describe, expect, it } from "vitest";
 import {
-  ACCEPTED_KEY_NAMES,
-  encodeKey,
   planSend,
   rejectUnknownSendFlags,
   resolveSendInput,
@@ -12,76 +9,6 @@ import {
 
 const START = "\x1b[200~";
 const END = "\x1b[201~";
-
-describe("encodeKey — named keys and modifier chords", () => {
-  it("maps named keys case-insensitively to their bytes", () => {
-    expect(encodeKey("Enter")).toBe("\r");
-    expect(encodeKey("return")).toBe("\r");
-    expect(encodeKey("Escape")).toBe("\x1b");
-    expect(encodeKey("esc")).toBe("\x1b");
-    expect(encodeKey("Tab")).toBe("\t");
-    expect(encodeKey("Space")).toBe(" ");
-    expect(encodeKey("Backspace")).toBe("\x7f");
-  });
-
-  it("uses the normal-cursor (CSI) form for arrows", () => {
-    expect(encodeKey("Up")).toBe("\x1b[A");
-    expect(encodeKey("down")).toBe("\x1b[B");
-    expect(encodeKey("Right")).toBe("\x1b[C");
-    expect(encodeKey("LEFT")).toBe("\x1b[D");
-    expect(encodeKey("Home")).toBe("\x1b[H");
-    expect(encodeKey("End")).toBe("\x1b[F");
-  });
-
-  it("folds C-<char> chords to their control byte", () => {
-    expect(encodeKey("C-c")).toBe("\x03");
-    expect(encodeKey("C-a")).toBe("\x01");
-    expect(encodeKey("c-z")).toBe("\x1a"); // case-insensitive prefix + letter
-    expect(encodeKey("C-[")).toBe("\x1b"); // 0x5b & 0x1f = 0x1b
-    expect(encodeKey("C-space")).toBeUndefined(); // only a single char after C-
-    expect(encodeKey("C- ")).toBe("\x00"); // Ctrl+Space → NUL
-  });
-
-  it("prefixes ESC for M-<char> (meta/alt), char verbatim", () => {
-    expect(encodeKey("M-b")).toBe("\x1bb");
-    expect(encodeKey("m-B")).toBe("\x1bB");
-    expect(encodeKey("M-.")).toBe("\x1b.");
-  });
-
-  it("returns undefined for unknown names and unmappable chords", () => {
-    expect(encodeKey("Foo")).toBeUndefined();
-    expect(encodeKey("C-1")).toBeUndefined(); // digit has no control byte
-    expect(encodeKey("")).toBeUndefined();
-    expect(encodeKey("C-")).toBeUndefined(); // no char after the chord prefix
-  });
-});
-
-describe("ACCEPTED_KEY_NAMES — the help vocabulary stays in lockstep with the table", () => {
-  // Split the human string into individual tokens: comma-separated, with the
-  // arrow cluster joined by slashes (`Up/Down/Left/Right`).
-  const tokens = ACCEPTED_KEY_NAMES.split(",")
-    .flatMap((s) => s.split("/"))
-    .map((s) => s.trim())
-    .filter((s) => s.length > 0);
-
-  it("every advertised name resolves via encodeKey", () => {
-    for (const name of tokens) {
-      expect(encodeKey(name), `advertised key "${name}" must encode`).not.toBe(
-        undefined,
-      );
-    }
-  });
-
-  it("reaches every byte in NAMED_KEY_BYTES (a new key can't drift the help)", () => {
-    const advertised = new Set(tokens.map((t) => encodeKey(t)));
-    for (const [name, bytes] of Object.entries(NAMED_KEY_BYTES)) {
-      expect(
-        advertised.has(bytes),
-        `key "${name}" (${JSON.stringify(bytes)}) is in NAMED_KEY_BYTES but not reachable from ACCEPTED_KEY_NAMES — add it`,
-      ).toBe(true);
-    }
-  });
-});
 
 describe("resolveSendInput — the arg-legality matrix", () => {
   const base = {
